添加鼠标事件处理支持，创建 MouseCodes.h 文件.将引擎调用从dll转为了lib

+++ Currsor/src/Currsor/Imgui/ImguiLayer.cpp
@@ -83,7 +83,8 @@ namespace Currsor
 		*/
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()),
+		                        static_cast<float>(app.GetWindow().GetHeight()));
 
 		float time = static_cast<float>(glfwGetTime());
 		io.DeltaTime = m_Time > 0.0f ? (time - m_Time) : (1.0f / 60.0f);
@@ -167,7 +168,7 @@ namespace Currsor
 	bool ImGuiLayer::OnWindowResizeEvent(WindowResizeEvent& e)
 	{
 		ImGuiIO& io = ImGui::GetIO();
-		io.DisplaySize = ImVec2(e.GetWidth(), e.GetHeight());
+		io.DisplaySize = ImVec2(static_cast<float>(e.GetWidth()), static_cast<float>(e.GetHeight()));
 		io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
 		glViewport(0, 0, e.GetWidth(), e.GetHeight());
